Rename all files with toctrees to .trst
To be able to use files from the blueprint in other documentation
projects, those files can not be used in any toctree in the blueprint.
This is because Sphinx parses all the files in all directories and
looks for toctrees. If it finds one it tries to create the sidebar
from them. If one file is found in two seperate toctrees at the same
time Sphinx gets confused and crashes leaving a blank sidebar.

In the conf.py file we can set up which suffixes we want Sphinx
to traverse to find all files. This patch atts the suffix .trst
to the list and renames all the files which have a toctree to
.trst too. This way the blueprint can be build on itself with the
.trst files, but the real project doesn't pick them up because it
only has .rst in their list.