Jenkinsfile: test builds with latest SWF baseline
Add testing of the BP together with the latest PELUX SWF baseline.
This ensures that any changes in the BP that breaks the baseline are caught
before merging changes in the BP. This way, when a change is merged in the
BP without issues, we could safely bump the BP submodule in the baseline.

Signed-off-by: Sapan <SKumarDas@luxoft.com>