Moving vert-config from swf to swf-blueprint
Previously software factory blueprint was not building because of files
being included on swf instead of swf-blueprint. I have moved that file
from swf to software-factory-blueprint.

Also had to add `:orphan:` which allows for a file to not be included in
a toctree (which is neccesary).

Because I added that file to the swf-blueprint, spellcheck wordlist had
to be updated as well.

Signed-off-by: Fisnik Hajredini <fhajredini@luxoft.com>